Remove proc evaluation in payload values.
We needed it to evaluate the person data proc, that it's defined as part of the scope in the Rails middleware. Because it's defined before the person data is set by the controllers, we still need to use a proc for it. We evaluate the person data while we are building the payload but removing the evaluation of all defined proc in the configuration.

We'll think in the future about the correct feature for the users, but this approach was a little dangerous.