Fix Rollbar error caused when ActiveRecord was present, but not used.
If the ActiveRecord gem was required in a Rails app, but not actually
used or configured (in favor of some other ORM or no ORM), then Rollbar
would internally error whenever it tried to send an exception message.
This was being caused by the person tracking functionality seeing that
ActiveRecord existed, but then it would raise a
ActiveRecord::ConnectionNotEstablished error when it actually tried to
use the connection. This in turn led to a rather non-obvious error in
the logs:

```
[Rollbar] Reporting internal error encountered while sending data to Rollbar.
[Rollbar] Sending payload
[Rollbar] Got unexpected status code from Rollbar api: 422
[Rollbar] Response: {
  "err": 1,
  "message": "Invalid format. data.context object value found, but a string is required."
}
```

This should fix this person-tracking situation by first checking whether
ActiveRecord is actually connected before trying to do
ActiveRecord-specific functions.