Fix when MultiJson is using adapters other than Oj
Whenever Yajl is used for MultiJson, before this patch,
Rollbar would find ::Yajl instead of Rollbar::JSON::Yajl,
and ::Yajl does exist but it does not respond to `options`,
causing failures.

It's surprising that Rollbar::JSON.const_get(:Yajl) would
find ::Yajl instead, it's like using Rollbar::JSON::Yajl to
refer to ::Yajl, this is really unfortunate for Ruby.
But given current status, we cannot rely on const_get in
this case. Therefore we'll need to search on constants
instead.

This has another advantage -- we no longer need to rescue,
saving an exception.