Notify deploy only on primary :rollbar_role server
Hi, I've seen that on every deploy on a multi app node the deploy gets noticed once for each node which get the deploy, so on a tree app node connected to rollbar there's three notifications of the same deploy which makes tree notices on a slack channel if you have slack and rollbar paired.

So perhaps it's a good idea notify the deploy only on primary server.

*Important note*: I've made this PR using the github editor and my knowledge of capistrano so it is not properly tested. I planed only to open an issue but the change is so small that I've created it as PR.