Add Rubocop config
Since we are still supporting Ruby 1.8, I've added disabled some
settings so we don't have errors on them.

LineLength is set to 1000 for now, since we have lot of long lines. Once
we do a first work fixing the style we can then enable it and be focus
on long lines

Style/DoubleNegation is disabled as a personal preference :D

vendor/** and gemfiles/vendor** are excluded

[branch #2981]