Fix URL scrubbing and change to a functional object
After one scrubbing refactor we starting not passing the scrubing
configuration to the URL scrubber. This is fixed with this PR.

Also, since params scrubbers was designed as a functional object, URL
scrubber is not also a functional object without state. Just to maintain
the style for them.

Some tests have been added to the scrubbers and the request data
extractor module