Fixed broken JS plugin initialization.
When secure_headers gem is enabled, the rollbar.js plugin doesn't
initialize (the script tag is not getting inserted). I'm not familiar
with the rollbar plugin flow, but my expectation is, since the plugin is
loaded via Railtie initializer (not directly), the `plugin.execute`
queue is not emptied at this point (the `load!` is not fired anymore).

What I did, i moved the queue initialization do the `call` endpoint, and
passed the actual rollbar.js initialization body the the Railtie
initializer (which is invoked only when secure_headers are found).