Enable per worker error notification thresholds
It is useful to be able to tune the error reporting depending on the
job. You may have a job with inconsistent failure rates or one that is
critical to be notified if it fails 3x while 5 is a good general default
for your app. It's a simple change so supporting this seemed worthwhile