Load rails-rollbar.js plugin always with Plugin#execute
We have two different cases when loading rails-rollbar.js plugin, having
SecureHeaders::Middleware defined or not.

In both cases we will check if Rollbar.configuration.js_enabled is true
or not. Since that flag is set by the user using Rollbar.configure
block, we need to delay the execution of the plugin code until the last
moment.

We do that using Rollbar::Plugin#execute. All the plugins loaded in that
way are sure to be loaded having a correct Rollbar.configuration.