Do big changes in configuration behavior
In order to make the gem code viable for the "temporaly override
notifier configuration" feature, we needed to do few big changes in the
behavior.

Before this commit, Rollbar.configuration is an object that doesn't
mutate (at first), and its values are propagated to the notifier,
Rollbar.notifier.

This presents some problems if we want to override the config per
notifier at certain moment, since we are using Rollbar.configuration at
some places, so, it's a global value that is not part of the notifier.

With this commit, Rollbar.configuration points to
Rollbar.notifier.configuration, so we don't have a global configuration
without owner and we always use the current thread notifier's
configuration.

Other changes:

- Rollbar.reset_notifier! doesn't clear the notifier but clears the
  scope object for the notifier. The notifier is the owner of the
  configuration, so we cannot drop it. But we ensure that the scope is
  fresh when using this method. This is the current behavior that the
  middleware expect for this method
- Rollbar.configure, Rollbar.reconfigure and Rollbar.preconfigure
  delegate to Rollbar.notifier.

At the end, we ensure that Rollbar.configuration is *always* the
configuration for the notifier being used by the current thread. Having
this scenario, we are not able to create new temporaly notifiers with
new configuration and return to the previous one when a block code is executed