Add Rollbar::Notifier#with_config
Many users ask how to override the configuration for a specific block of
code. This PR adds that functionality through the method
`Rollbar.with_config`, that delegates to the per-thread
Rollbar::Notifier instance.

The argument passed to `#with_config` is a `Hash` having key names equal
to the allowed configuration options and the new value to use for the
given block of code.

Example:

```ruby
Rollar.configure do |config|
  config.environment = 'production'
end

Rollbar.with_config(environment: 'staging') do
  begin
    # do something that will crash
  rescue => e
    Rollbar.error(e)
  end
end
```

In the example from above, the reported error will use `staging` as
environment instead of `production`.