Send correct route params
Rails env['action_dispatch.request.parameters'] include all the post,
and json received data. Our API says that `params` are the route params,
and we are sending all the get, post, json data on it.

Since we are sending that information using the correct GET, POST and
body keys, we should just send the route params in `params` key.

The route params will be, :controller, :action, :format (if used) and
any route param in the given URL.