Ensure params filtering take precedence over hash/array objects
This matches the current Rails implementation of parameter filtering,
where it is possible to set a filter that will hide an entire array or
nested hash, instead of just the inner keys.

Consider the following params payload:

    user: { username: "foo", password: "bar" }

Currently we can set a filter for both "username" and "password", but
setting for the entire "user" params hash does not work, it's simply
ignored, because we parse hashes separately looking at each individual
key first.

With this change, setting a filter for "user" would effectively hide it
from the payload, scrubbing the field as expected.