Handle malformed JSON request bodies
If someone makes a request with a body and a JSON content-type,
we attempt to parse it as part of the request data extraction process.
If the body is actually not well formed JSON, then currently we just
throw the whole thing away and return an empty object. This change
instead returns the raw body as a string.

Fixes #618