Don't send person's email address as a default
Context:

Sending a user's email address to a third-party is not desirable under The EU
General Data Protection Regulation (GDPR) and sending it to Rollbar is not a
sensible default.

Change:

The default is *not* to send the email address but how to do so remains a
documented option.

Specs for EU GDPR Default